refactor: winningNumber 검증 테스트를 LottoTest로 이동
