+++ src/test/java/lotto/model/LottoTest.java
@@ -5,8 +5,8 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 class LottoTest {
     @Test
@@ -38,4 +38,39 @@ class LottoTest {
         assertThat(actual).isTrue();
     }
 
+    @Test
+    @DisplayName("로또 번호 개수가 6개 미만인 경우 예외 처리")
+    void validateNumberOfWinningNumbersUnder6Test() {
+        assertThatThrownBy(() -> {
+            new Lotto(Arrays.asList(1, 2, 3, 4, 5));
+        }).isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("로또 번호 개수는 6개로 입력해주세요.");
+    }
+
+    @Test
+    @DisplayName("로또 번호 개수가 6개 초과인 경우 예외 처리")
+    void validateNumberOfWinningNumbersOver6Test() {
+        assertThatThrownBy(() -> {
+            new Lotto(Arrays.asList(1, 2, 3, 4, 5, 6, 7));
+        }).isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("로또 번호 개수는 6개로 입력해주세요.");
+    }
+
+    @Test
+    @DisplayName("볼 번호의 범위가 1~45가 아닌 경우 예외 처리")
+    void validateRangeWinningNumbersTest() {
+        assertThatThrownBy(() -> {
+            new Lotto(Arrays.asList(1, 2, 3, 4, 55, 6));
+        }).isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("볼 번호가 1~45 범위 내에 해당하지 않습니다.");
+    }
+
+    @Test
+    @DisplayName("로또 번호 중 중복된 값이 입력될 경우 예외 처리")
+    void validateDuplicationWinningNumbersTest() {
+        assertThatThrownBy(() -> {
+            new Lotto(Arrays.asList(1, 2, 3, 3, 4, 6));
+        }).isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining("로또 번호에 중복이 존재합니다.");
+    }
 }
